tools: Improve assets.sh script

+++ assets.sh
@@ -2,9 +2,16 @@
 #
 # SPDX-License-Identifier: GPL-3.0-only
 #
-# Copyright (c) 2021, Antonio Niño Díaz
+# Copyright (c) 2021-2022, Antonio Niño Díaz
 
 set -e
+
+# Paths to tools
+
+export SUPERFAMICONV=${PWD}/SuperFamiconv/bin/superfamiconv
+export PNGS2STRIP=${PWD}/tools/build/pngs2strip/pngs2strip
+export UMOD_PACKER=${PWD}/umod-player/build/packer/umod_packer
+export BIN2C=${PWD}/tools/build/bin2c/bin2c
 
 # Build tools
 
@@ -14,22 +21,19 @@ cmake ..
 make
 popd
 
-pushd SuperFamiconv
-rm -rf build ; mkdir build ; cd build
-cmake ..
-make -j`nproc`
-popd
+if [ ! -d "${SUPERFAMICONV}" ]; then
+    pushd SuperFamiconv
+    make -j`nproc`
+    popd
+fi
 
-pushd umod-player
-rm -rf build ; mkdir build ; cd build
-cmake ..
-make
-popd
-
-# Paths to tools
-
-export SUPERFAMICONV=SuperFamiconv/build/superfamiconv
-export PNGS2STRIP=tools/build/pngs2strip/pngs2strip
+if [ ! -d "${UMOD_PACKER}" ]; then
+    pushd umod-player
+    rm -rf build ; mkdir build ; cd build
+    cmake ..
+    make
+    popd
+fi
 
 # Prepare destination folder
 
@@ -89,7 +93,7 @@ bash sprites/minimap_menu_gbc/convert.sh ${OUT_DIR_SPRITES_MINIMAP_MENU_GBC}
 
 mkdir ${OUT_DIR}/audio
 
-umod-player/build/packer/umod_packer \
+${UMOD_PACKER} \
     ${OUT_DIR}/audio/umod_pack.bin \
     ${OUT_DIR}/audio/umod_pack_info.h \
     audio/songs/*/*.mod \
@@ -101,7 +105,7 @@ for dir in $(find built_assets -type d)
 do
     for f in $(find $dir -maxdepth 1 -iname *.bin)
     do
-        tools/build/bin2c/bin2c $f "$dir"
+        ${BIN2C} $f "$dir"
     done
 done
